requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
boto3==1.35.6
python-dotenv==1.0.1
pytest==8.3.3
pytest-django==4.9.0
gunicorn==23.0.0
python-decouple==3.8